server: split client, room, system into separate go files
